feat(init project): working in migrate to gitlab pr
