+++ .gitignore
@@ -71,3 +71,4 @@ dist/
 !.vscode/extensions.json
 
 # End of https://www.gitignore.io/api/node,visualstudiocode
+
