+++ README.md
@@ -1,6 +1,5 @@
 # danger-plugin-slack
 
-[![Build Status](https://travis-ci.org/julon/danger-plugin-slack.svg?branch=master)](https://travis-ci.org/julon/danger-plugin-slack)
 [![npm version](https://badge.fury.io/js/danger-plugin-slack.svg)](https://badge.fury.io/js/danger-plugin-slack)
 [![semantic-release](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--release-e10079.svg)](https://github.com/semantic-release/semantic-release)
 [![Greenkeeper badge](https://badges.greenkeeper.io/julon/danger-plugin-slack.svg)](https://greenkeeper.io/)
@@ -12,14 +11,14 @@
 Install:
 
 ```sh
-yarn add danger-plugin-slack --dev
+yarn add danger-plugin-gitlab-slack --dev
 ```
 
 To send the Danger report to slack:
 
 ```js
 // dangerfile.js
-import slack from 'danger-plugin-slack'
+import slack from 'danger-plugin-gitlab-slack'
 
 const options = {
   webhookUrl: "YOUR_SLACK_WEBHOOK_URL" // only the webhook URL is required
