+++ package.json
@@ -1,30 +1,29 @@
 {
-  "name": "danger-plugin-slack",
+  "name": "danger-plugin-gitlab-slack",
   "description": "Report to Slack the result of Danger",
   "author": {
-    "name": "Julon Lou",
-    "email": "contact@julonlou.com"
+    "name": "unfernandito",
+    "email": "lfbu.112@gmail.com"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/julon/danger-plugin-slack.git"
+    "url": "https://github.com/unfernandito/danger-plugin-gitlab-slack.git"
   },
   "bugs": {
-    "url": "https://github.com/julon/danger-plugin-slack/issues"
+    "url": "https://github.com/unfernandito/danger-plugin-gitlab-slack/issues"
   },
-  "homepage": "https://github.com/julon/danger-plugin-slack#readme",
+  "homepage": "https://github.com/unfernandito/danger-plugin-gitlab-slack#readme",
   "keywords": [
     "danger",
     "danger-plugin",
     "slack"
   ],
-  "version": "1.1.2",
+  "version": "0.0.0-development",
   "main": "dist/index.js",
   "types": "types/index.d.ts",
   "scripts": {
-    "precommit": "lint-staged",
-    "commit": "git-cz",
-    "commitmsg": "validate-commit-msg",
+    "lint-staged": "lint-staged",
+    "precommit": "npm run lint-staged && npm run test && npm run build",
     "build": "tsc",
     "test": "jest",
     "predocs": "rm -rf docs/",
@@ -34,33 +33,31 @@
     "prettier": "prettier",
     "prettier-write": "npm run prettier -- --parser typescript --no-semi --trailing-comma es5 --write --print-width 120",
     "prettier-project": "npm run prettier-write -- 'src/**/*.{ts,tsx}'",
-    "lint": "tslint \"src/**/*.ts\"",
-    "travis-deploy-once": "travis-deploy-once"
+    "lint": "eslint src/ --ext ts",
+    "prepare": "husky install"
   },
   "license": "MIT",
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=10.0.0"
   },
   "devDependencies": {
-    "@semantic-release/changelog": "^1.0.0",
-    "@semantic-release/git": "^3.0.0",
-    "@semantic-release/github": "^4.0.0",
-    "@semantic-release/npm": "^3.0.0",
-    "@slack/client": "^3.10.0",
-    "@types/jest": "^22.1.0",
-    "commitizen": "^2.9.6",
-    "cz-conventional-changelog": "^2.0.0",
-    "danger": "^3.0.5",
-    "husky": "^0.14.3",
-    "jest": "^22.0.3",
-    "lint-staged": "^6.0.0",
-    "prettier": "^1.10.2",
-    "semantic-release": "^12.2.2",
-    "travis-deploy-once": "^4.3.1",
-    "ts-jest": "^22.0.0",
-    "tslint": "^5.9.1",
-    "typescript": "^2.3.2",
-    "validate-commit-msg": "^2.12.1"
+    "@semantic-release/changelog": "5.0.1",
+    "@semantic-release/git": "9.0.0",
+    "@semantic-release/github": "7.2.3",
+    "@semantic-release/npm": "7.1.3",
+    "@slack/webhook": "6.0.0",
+    "@types/jest": "26.0.23",
+    "@typescript-eslint/eslint-plugin": "4.28.0",
+    "@typescript-eslint/parser": "4.28.0",
+    "danger": "10.6.4",
+    "eslint": "7.29.0",
+    "husky": "6.0.0",
+    "jest": "27.0.5",
+    "lint-staged": "11.0.0",
+    "prettier": "2.3.1",
+    "semantic-release": "^17.4.4",
+    "ts-jest": "27.0.3",
+    "typescript": "4.3.4"
   },
   "optionalDependencies": {
     "esdoc": "^1.0.1"
@@ -72,7 +69,7 @@
   },
   "lint-staged": {
     "*.@(ts|tsx)": [
-      "tslint --fix",
+      "npm run lint -- --fix",
       "npm run prettier-write --",
       "git add"
     ]
@@ -84,7 +81,7 @@
       "js"
     ],
     "transform": {
-      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+      ".(ts|tsx)": "ts-jest"
     },
     "testRegex": "(.test)\\.(ts|tsx)$",
     "testPathIgnorePatterns": [
