+++ src/utils/createMessage.ts
@@ -1,0 +1,72 @@
+import { IncomingWebhookSendArguments } from "@slack/webhook"
+import { GitHubPRDSL, DangerResults } from "danger"
+import { SlackOptions } from "../types"
+import { createAttachment } from "./createAttachment"
+import { createMarkdownAttachment } from "./createMarkdownAttachment"
+import { getDynamicEmoji } from "./getDynamicEmoji"
+
+export function createMessage(
+  pr: Partial<GitHubPRDSL> | GitHubPRDSL,
+  resultLists: DangerResults,
+  options: SlackOptions
+): IncomingWebhookSendArguments {
+  const msg: IncomingWebhookSendArguments = {
+    text: "",
+    username: options.username || "DangerJS",
+    icon_emoji: options.iconEmoji || ":open_mouth:",
+    attachments: [],
+  }
+
+  // custom iconUrl if set
+  if (options.iconUrl) {
+    msg.icon_url = options.iconUrl
+  }
+
+  // custom channel if set
+  if (options.channel) {
+    msg.channel = options.channel
+  }
+
+  // send only a custom text
+  if (options.text) {
+    msg.text = `${options.text}`
+  } else {
+    // send only the report
+
+    const fails = resultLists.fails
+    const warnings = resultLists.warnings
+    const messages = resultLists.messages
+    const markdowns = resultLists.markdowns
+
+    const prInfo = pr ? `<${pr.html_url}|*PR#${pr.number}* - ${pr.title}>` : "not info founded"
+    const prAuthor = pr ? `<${pr.user?.avatar_url}|${pr.user?.login}>` : "not author found"
+
+    msg.icon_emoji = getDynamicEmoji(fails, warnings)
+    msg.text = `${prInfo} by ${prAuthor}\n${pr.body}`
+
+    // add violations as slack attachments
+    if (!fails.length && !warnings.length && !messages.length) {
+      msg.text += "\nNo output to show."
+    } else {
+      if (msg.attachments) {
+        if (fails.length > 0) {
+          msg.attachments.push(createAttachment("Fails", "danger", fails))
+        }
+
+        if (warnings.length > 0) {
+          msg.attachments.push(createAttachment("Warnings", "warning", warnings))
+        }
+
+        if (messages.length > 0) {
+          msg.attachments.push(createAttachment("Messages", "#999", messages))
+        }
+
+        if (markdowns.length > 0) {
+          msg.attachments.push(createMarkdownAttachment("Comments", "#EEE", markdowns))
+        }
+      }
+    }
+  }
+
+  return msg
+}
