+++ tsconfig.json
@@ -1,23 +1,15 @@
 {
-    "compilerOptions": {
-        "target": "es6",
-        "module": "commonjs",
-        "rootDir": "src",
-        "outDir": "dist",
-        "allowJs": false,
-        "pretty": true,
-        "strictNullChecks": true,
-        "declaration": true
-    },
-    "lib":["es2017"],
-    "include": [
-        "src/**/*.ts",
-        "src/**/*.tsx",
-        "dangerfile.ts"
-    ],
-    "exclude": [
-        "dangerfile.ts",
-        "node_modules",
-        "dist"
-    ]
+  "compilerOptions": {
+    "target": "es6",
+    "module": "commonjs",
+    "rootDir": "src",
+    "outDir": "dist",
+    "allowJs": false,
+    "pretty": true,
+    "strictNullChecks": true,
+    "declaration": true
+  },
+  "lib": ["es2017"],
+  "include": ["src/**/*.ts", "src/**/*.tsx", "dangerfile.ts"],
+  "exclude": ["dangerfile.ts", "node_modules", "dist"]
 }
